+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [0.7.0] - 2023-08-06
+
+### Changed
+
+- Added claims phone_number and address
+- Added possibility to add claim picture with data url to id token or user info dependent in app settings.
+- Dependency updates
 
 ## [0.6.2] - 2023-07-28